feat(tests): amigrate action menu specs to ts
