+++ tests/action-menu/copy-link.spec.ts-snapshots/linux/chrome/2-board-link.png
