+++ tests/action-menu/copy-link.spec.ts-snapshots/linux/firefox/2-board-link.png
