+++ tests/action-menu/copy-link.spec.ts-snapshots/win32/chrome/2-board-link.png
