+++ tests/action-menu/copy-link.spec.ts-snapshots/win32/firefox/2-board-link.png
